pyproject.toml
[tool.poetry]
name = "longest-word"
version = "0.1.0"
description = ""
authors = ["trehan christie <trehan2010@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^8.3.3"
pylint = "*"
pyenchant = "^3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
